# pkg/restClient/types.go
// Package restClient provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.12.0 DO NOT EDIT.
package restClient

// Metadata Metadata represents a single value of metadata.
type Metadata struct {
	Key   string `json:"key"`
	Value string `json:"value"`
}

// MetadataList defines model for MetadataList.
type MetadataList struct {
	Metadata []Metadata `json:"metadata"`
}

// MetadataResponse defines model for MetadataResponse.
type MetadataResponse struct {
	Metadata []StoredMetadata `json:"metadata"`
}

// StoredMetadata StoredMetadata represents all stored metadata values for a given key.
type StoredMetadata struct {
	Key    string   `json:"key"`
	Values []string `json:"values"`
}

// MetadataServiceDeleteParams defines parameters for MetadataServiceDelete.
type MetadataServiceDeleteParams struct {
	Key   *string `form:"key,omitempty" json:"key,omitempty"`
	Value *string `form:"value,omitempty" json:"value,omitempty"`
}

// MetadataServiceCreateOrUpdateMetadataJSONRequestBody defines body for MetadataServiceCreateOrUpdateMetadata for application/json ContentType.
type MetadataServiceCreateOrUpdateMetadataJSONRequestBody = MetadataList